[#3558] ParenthesesCorrector extraction
